% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_client.R
\name{mod_client_ui}
\alias{mod_client_ui}
\title{mod_client_ui and mod_client_server}
\usage{
mod_client_ui(id)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}
}
\value{
a \code{shiny::\link[shiny]{tagList}} containing UI elements
}
\description{
A shiny Module.
}
\details{
The UI portion of the module for displaying the Client datatable
}
