% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ticket.R
\name{mod_ticket_ui}
\alias{mod_ticket_ui}
\title{mod_ticket_ui and mod_ticket_server}
\usage{
mod_ticket_ui(id)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}
}
\value{
a \code{shiny::\link[shiny]{tagList}} containing UI elements
}
\description{
A shiny Module.
}
\details{
The UI portion of the module for displaying the ticket datatable
}
